Final prelim code

+++ 4191_g6/lesgo_g6.cydsn/CircularQ.c
@@ -22,20 +22,20 @@ CircularQ circularq_new() {
     };
 }
 
-void enqueue(volatile CircularQ* q, uint16 elem) {
+void enqueue(volatile CircularQ* q, float elem) {
     CIRCULARQ_IDX_INC(q->head);
     q->buf[q->head] = elem;
 }
 
 
-uint16 dequeue(volatile CircularQ* q) {
-    uint16 elem = q->buf[q->tail];
+float dequeue(volatile CircularQ* q) {
+    float elem = q->buf[q->tail];
     CIRCULARQ_IDX_INC(q->tail);
     return elem;
 }
 
 
-uint16 peek(volatile CircularQ* q) {
+float peek(volatile CircularQ* q) {
     return q->buf[q->tail];
 }
