+++ 4191_g6/lesgo_g6.cydsn/CircularQ.h
@@ -21,14 +21,14 @@
 typedef struct CircularQ {
     int8 head;
     int8 tail;
-    uint16 buf[CIRCULARQ_LEN];
+    float buf[CIRCULARQ_LEN];
 } CircularQ;
 
 
 CircularQ circularq_new();
-void enqueue(volatile CircularQ* q, uint16 elem);
-uint16 dequeue(volatile CircularQ* q);
-uint16 peek(volatile CircularQ* q);
+void enqueue(volatile CircularQ* q, float elem);
+float dequeue(volatile CircularQ* q);
+float peek(volatile CircularQ* q);
 
 
 #endif  // CIRCULARQ_H
