+++ 4191_g6/lesgo_g6.cydsn/ultrasonic.h
@@ -25,10 +25,10 @@ CY_ISR_PROTO(usr_fr_isr);
 
 // API
 void ultrasonic_setup(void);
-uint16 us_l_get_dist(void);
-uint16 us_r_get_dist(void);
-uint16 us_fl_get_dist(void);
-uint16 us_fr_get_dist(void);
+float us_l_get_dist(void);
+float us_r_get_dist(void);
+float us_fl_get_dist(void);
+float us_fr_get_dist(void);
